apps/slinky: use led when available
Not all targets may have LED pins defined and available.
Use LED related functionality when available.
